backend/requirements.txt
fastapi
uvicorn
sqlmodel
sqlalchemy
psycopg2-binary
python-dotenv 